+++ Code/profile.cpp
@@ -1,7 +1,7 @@
 #include "profile.h"
 #include <iostream>
 
-Profile::Profile(int basalRate, int carbohydrateRatio, int correctionFactor, int targetGlucoseLevel, string profileName)
+Profile::Profile(double basalRate, int carbohydrateRatio, int correctionFactor, int targetGlucoseLevel, string profileName)
 { // im sorry eli, but this is easir to read (for me at least)
     _basalRate = basalRate;
     _carbohydrateRatio = carbohydrateRatio;
